feat: Only core modules can use all language features (#20)
While it's convenient for implementation of the language itself to have all the tools available, newer languages like Elm and Roc have shown how valueble it can be to restrict "user space" features to a specific subset.

This is also nice for experimentation behind the scenes without having messy language feature extensions like i.e. Haskell does.

The parser looks at the base namespace of the module declaration and makes all language features reachable when it starts with "core". This could change later but serves its purpose for now.