feat(lib): Detect and setup application main so the module graph builds itself naturally via import calls (#65)
Instead of manually sorting modules we just detect the application
module, capture its main function and stash it in a thunk so that when
the app is actually initialized by the user it builds the module graph
on its own via import calls.

We're still not generating useful code and the capabilities haven't been
integrated yet.