style(core): Play around with core library structure but dismiss "module as directory" for now (#55) (#71)
Having modules just be comprised of files in the same directory reduces
some boilerplate and it probably reduces the amount of files you
actually need due to cyclic data dependencies not being a thing but the
approach doesn't seem to fit well with the language goal of favoring
things being explicit. It also makes compile time module configs a lot
more difficult because it would be non-local to the current file

Not ruling out trying this some more but for now we'll stay inside the
ML tradition of one module per file.