style(core): Play around with higher-kinded types (#55) (#79)
We want to be able to express the necessary set theory maths for
"proper" operator overloading so we could expose that functionality in a
safe way. With type concept contracts we could even fuzz test compliance
with expected behavior. This is mainly for independently-versioned
library-supplied not-mathematically-a-number types like F32/F64/F128 we
don't want inside the language core.